requirements.txt
Django==1.10
django-braces==1.9.0
django-oauth-toolkit==0.10.0
djangorestframework==3.4.5
mysqlclient==1.3.7
oauthlib==1.0.3
six==1.10.0
